+++ Sources/PeripheryKit/Results/Baseline.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+/// A baseline set of declarations that are excluded from results.
+public enum Baseline: Codable {
+    case v1(usrs: [String])
+
+    public var usrs: Set<String> {
+        switch self {
+        case .v1(let usrs):
+            return Set(usrs)
+        }
+    }
+}